standalone_app/requirements.txt
customtkinter
Pillow
opencv-python
numpy
pyinstaller
